% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arima.R
\name{arima_process}
\alias{arima_process}
\title{ARIMA processes}
\usage{
arima_process(
  .distfn = "rnorm",
  n = NULL,
  y0 = 0,
  P = 1,
  D = 0,
  Q = 0,
  beta = 0.5,
  theta = 0.5,
  ...
)
}
\arguments{
\item{.distfn}{What function should be used to draw the innovations from? 
Must be a string. 
Defaults to `"rnorm"`, should work with any `r_` compatible function.}

\item{n}{sample size, only has to be specified if function is called 
outside of a tibble or similar data context}

\item{y0}{The initial condition}

\item{P}{If an integer, then adds lags from 1 to P, if a numeric vector,
then only given lags.}

\item{D}{If an integer, then adds lags from 1 to P, if a numeric vector,
then only given lags.}

\item{Q}{If an integer, then adds lags from 1 to P, if a numeric vector,
then only given lags.}

\item{beta}{autoregressive parameters}

\item{theta}{moving average parameters}

\item{...}{arguments to be passed to `.distfn`}
}
\description{
ARIMA processes
}
