% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/white_noise.R
\name{white_noise}
\alias{white_noise}
\title{White Noise}
\usage{
white_noise(.distfn = "rnorm", n = NULL, y0 = 0, ...)
}
\arguments{
\item{.distfn}{What function should be used to draw the innovations from? Must be a string. 
Defaults to `"rnorm"`, should work with any `r_` compatible function.}

\item{n}{sample size, only has to be specified if function is called outside of a tibble or similar data context}

\item{y0}{The initial condition}

\item{...}{arguments to be passed to `.distfn`}
}
\description{
White Noise
}
\examples{

white_noise(n = 10, sd = 2)

white_noise("runif", n = 10, max = 2)


}
